Price_Estimator/Price_Estimator_Backend/requirements.txt
streamlit
pandas
xlsxwriter
streamlit-autorefresh>=0.0.2
fastapi
uvicorn
pydantic
openai

python-dotenv